+++ articles/views/__init__.py
@@ -1,2 +1,2 @@
 from .views import ArticleDetailView
-from .lists import LatestArticlesView
+from .lists import LatestArticlesView, CategoryArticlesView