+++ articles/views/lists.py
@@ -2,10 +2,10 @@ from datetime import datetime
 
 from django.core.paginator import Paginator
 from django.http import HttpResponse
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 from django.views import View
 
-from articles.models import Article
+from articles.models import Article, Category
 
 DEFAULT_SHOWN = 10
 
@@ -17,6 +17,28 @@ class LatestArticlesView(View):
         shown_param = '&shown={0}'.format(shown) if shown != DEFAULT_SHOWN else ''
 
         article_list = Article.objects.select_related('author').all()\
+            .filter(publication_date__lte=datetime.now(), state__exact='PB')\
+            .order_by('-publication_date')
+        paginator = Paginator(article_list, shown)
+
+        articles = paginator.get_page(page)
+
+        context = {'latest_articles': articles, 'shown_param': shown_param}
+
+        html = render(request, 'articles/latest.html', context)
+
+        return HttpResponse(html)
+
+
+class CategoryArticlesView(View):
+    def get(self, request, slug=None):
+        category = get_object_or_404(Category, slug=slug)
+
+        page = request.GET.get('page')
+        shown = request.GET.get('shown', DEFAULT_SHOWN)
+        shown_param = '&shown={0}'.format(shown) if shown != DEFAULT_SHOWN else ''
+
+        article_list = category.articles.select_related('author').all()\
             .filter(publication_date__lte=datetime.now(), state__exact='PB')\
             .order_by('-publication_date')
         paginator = Paginator(article_list, shown)